Добавил класс User для хранения информации о пользователе Добавил протокол UserServiceProtocol с функцией проверки логина Выполнил задание
