+++ Navigation.xcodeproj/xcuserdata/maksimdonskih.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -4,6 +4,11 @@
 <dict>
 	<key>SchemeUserState</key>
 	<dict>
+		<key>Debug.xcscheme_^#shared#^_</key>
+		<dict>
+			<key>orderHint</key>
+			<integer>2</integer>
+		</dict>
 		<key>Navigation release.xcscheme</key>
 		<dict>
 			<key>orderHint</key>
@@ -14,15 +19,15 @@
 			<key>orderHint</key>
 			<integer>1</integer>
 		</dict>
-		<key>StorageService.xcscheme_^#shared#^_</key>
-		<dict>
-			<key>orderHint</key>
-			<integer>2</integer>
-		</dict>
 	</dict>
 	<key>SuppressBuildableAutocreation</key>
 	<dict>
 		<key>1659D08A2909CB53006843C6</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>16F1327E29143D530001D766</key>
 		<dict>
 			<key>primary</key>
 			<true/>
